WIP commit of adding new routes from the js stuff
I added a Endpoint.setup thing that does a lot of the repetitive work for you, it will probably be expanded. When using it you need to deconstruct the result in a form of ((avatar1,...,) base).

The base should be the name of the image and dir, with an optional extension. I may change this to automatically grab the images and save it in a cache on startup though. Idk, I'll figure something out. The point is the directory needs to be the same as the omage name so some of those will be changed.

I'm planning on moving what I can to the new self.setup style, but I doubt it'll work for all routes with no issues.